Move Ui class into it own package
