+++ src/test/java/seedu/cardcollector/command/CommandTest.java
@@ -1,7 +1,7 @@
 package seedu.cardcollector.command;
 
 import org.junit.jupiter.api.Test;
-import seedu.cardcollector.Ui;
+import seedu.cardcollector.ui.Ui;
 import seedu.cardcollector.UploadUndoState;
 import seedu.cardcollector.card.CardHistoryEntry;
 import seedu.cardcollector.card.CardHistoryType;
